Relocate all routes to main module __init__.py and add route test
